rename files for wider use, update footer format
