+++ src/Components/Core/Footer.css
@@ -3,4 +3,6 @@
   margin-top: auto;
   justify-content: center;
   background-color: #1E6A9C;
+  height: 65px;
+  align-items: center;
 }
