+++ src/Components/Core/Footer.js
@@ -1,4 +1,4 @@
-import footer from "../../images/footer.svg";
+import footer from "../../images/inverse-logo.svg";
 import { Link } from "react-router-dom";
 import "./Footer.css";
